Renamed class TrioSet (with geno only) to gTrio
The new class gTrio is a simplified version of the TrioSet object from MD that contains only genotype data.  The genotype data is stored as an array within the AssayData.  The AssayData is a slot inherited from the gSet defined in Biobase, which in turn is an extension of the classic eSet.  The original TrioSet extended the gSet to include also pedigree and father and mother phenotype data.  The phenotype data is more easily thought of as covariate data.  There are many features to the eSet and gSet that are not being used by gTrio.  gTrio is meant to be lightweight and simple.