Added TransCount method for GRangesList
Now TransCount can operate on a GRanges list and will return a vector of same length as the GRangesList.  This vevtor will contain the count of transmissions for each GRange element.

It has one loop -- the loop through the GRangesList.  This couls surely be made more efficient with lapply in parallel somehow.