Fixed TransCount() and Updated ScanTrio()
There was an error in TransCount().  We don't need to loop through ranges in the GRanges object.  We treat each GRanges object as one window.  We do not require that the window be one range, e.g., the 'outside of window' GRange object has two ranges.  We only want to loop through elements of the GRangesList.  This is a list of GRanges objects that are to be treated as windows.