Debugged some problems
Still having trouble with the M O switch.  I don't know why but for some reason the code only generates correct results when I switch the genotpye matrixes for M and O in the GenoTrio() object.  I have no idea why this is happening, but I don't think the problem is on the data level becuase plink and SnpStats behave correctly (seemingly).