Created generator function for CNVTrioExperiment
There have already been several methods available for ste that were need for cte.  I copied and pasted for a temporary patch.  This needs to be changed so that they both inherit from the same class, or functions exist in a functions.R file that are called by both methods.  The latter makes the structure very cryptic for the reader of the source code.