TrioAssay, cnv(), and reduce() in CNVMatrix()
Added the accessor method cnv() and renamed GenoTrio to TrioAssay.

CNVMatrix() was giving counts greater than one indicating that the subject overlapped with themselves because the component GRanges object is a non-overlapping partition.  Thus if anything were to overlap it would be between subjects, and in CNVMatrix we do one subject at a time.  So, there is a problem with self-overlap in the PennCNV data, which likely will lead to overcounts, and therefore false positives.