Added function p.min to functions.R
This function finds the minimum component-wise p-value for each region
and returns a DataFrame of regions.  It requires that the GRanges
object in the input DataFrame have name "grange".  An obvious
improvement is to make general both the set function, e.g.,
minimum, and the name of the Granges object, "grange".