Updated inherited-deletions vignette.
Now we count as transmitted all deletions that have a transmitted component anywhere in the region.  It takes an hour to count for all approximately two thousand regions.  Tomorrow I will use binom.test to compute the CIs for the transmission rate.

Note that a trio with trio-state 111 is ambiguousm as it is both a transmission and a non-transmission.  Currently it is counted towards transmitted and not towards untransmitted.