Updated Inherited Deletions Vignette.
On the web page there is now a link to the New Counting method within the CLP Association vignette.  Here we implement the change in counting and plot a histogram of overall tranmission rates per region, as well as the CIs per region , stratified by cleft, control, and all.

We still see regions of undertransmission.  If our hypothesis is that these are regions with excessive false positve rates, and that they should be eliminated from the study than we need a systematic way to categorize bad regions.

It still may be worthwhile to investigate correlation between marker density and under-transmission.  Some physical property, whether it is biology, or the chip design, is likely to be causing the under-transmission/false positives.