Updated Inherited Deletions vignette.
Now, we also see in the vignette the construction of the CNV matrices and FamilyExperiment objects.

We are relying heavily on caching within Asciidoc.  Most but, not all of the R code is evaluated.  The construction of the CNV matrix and counting transmissions by region are both comutationally intensive, taking about an hour each.

Note that the vignette could not be run from start to finish from scratch.  There are two instances where th epackage would beed to be reinstalled and reloaded.  This is due to saving data in the ./data/ directory at two steps.