A few small changes to inherited deletions vignette.
Cleaned up some whitespace.

Added a line to show the number of regions that are included in analysis.Note that right now regions with one component are not being considered.  It is highly unlikely that a region with over 50 observed deltions will only have one component.  Nonetheless there are a lot of them and we need to consider them.  This will be done soon, but it was a priority to get the figures finished for Terri's progress report.