Added ped2fe function to functions.R.
This function uses snpStats to read a pedigree and map file (plink format) and created the corresponding FamilyExperiment (fe) object.

As of now it reuqires that you include the number of SNPs in the ped (and map) file.  This shouldn't be so.  There seems to be a bug in snpStats:::read.pedfile.

Note also that Holger has a read.pedfile function in trio.  Perhaps this is better, but I am familiar with and have become to prefer the SnpMatrix class.