Include zero count subjects in CNV matrix
Previously if a subject had no deletions that satisfied the inclusion
requirements then that subject had no entry in the CNV matrix.  Thus
the trio-states probably weren't being formed with that subject as
normal copy number, but instead the trio-state was not even computed
as that subject was seen to be missing.